requirements.txt
Flask==2.3.1
psycopg2-binary==2.9.6
pandas==2.0.1
python-dotenv==1.0.0
pytest==7.3.1